// lib/api-client/models/CreateMediaDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateMediaDto = {
    url: string;
    type: CreateMediaDto.type;
};
export namespace CreateMediaDto {
    export enum type {
        IMAGE = 'image',
        VIDEO = 'video',
    }
}